requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
jinja2==3.1.3
python-dotenv==1.0.1
sqlalchemy==2.0.29
asyncpg==0.29.0
python-multipart==0.0.9
passlib
databases
bcrypt